requirements.txt
# FastAPI & Server
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Environment
python-dotenv==1.0.1

# Google Services
gspread==6.1.2
oauth2client==4.1.3
google-generativeai==0.8.3

# AI & ML
sentence-transformers==3.2.1
faiss-cpu==1.9.0
numpy==2.1.3

# Data Processing
pandas==2.2.3

# HTTP Client
httpx==0.27.2

# Utilities
python-multipart==0.0.12
